+++ apps/blog-app/src/components/layout/main-header.tsx
@@ -6,7 +6,7 @@ type Props = {
 
 export const MainHeader: FC<Props> = () => {
   return (
-    <header className="py-5 bg-purple-500 text-white text-center">
+    <header className="py-5 text-center text-white bg-purple-500">
       NextJs monorepo example - the blog-app
     </header>
   );