+++ apps/web-app/src/app-providers.tsx
@@ -1,8 +1,10 @@
 import { CacheProvider, EmotionCache } from '@emotion/react';
+import { ThemeProvider as MuiThemeProvider } from '@mui/material';
 import { FC } from 'react';
 
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { createEmotionCache } from '@/core/nextjs/create-emotion-cache';
+import { muiTheme } from '@/themes/mui/mui.theme';
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -23,9 +25,13 @@ export const AppProviders: FC<Props> = (props) => {
   const { emotionCache = clientSideEmotionCache } = props;
   return (
     <CacheProvider value={emotionCache}>
-      <QueryClientProvider client={queryClient}>
-        {props.children}
-      </QueryClientProvider>
+      <MuiThemeProvider theme={muiTheme}>
+        {/* Mui CssBaseline disabled in this example as tailwind provides its own */}
+        {/* <CssBaseline /> */}
+        <QueryClientProvider client={queryClient}>
+          {props.children}
+        </QueryClientProvider>
+      </MuiThemeProvider>
     </CacheProvider>
   );
 };