+++ packages/db-main-prisma/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @your-org/db-main-prisma
 
+## 2.7.0
+
+### Minor Changes
+
+- [#589](https://github.com/belgattitude/nextjs-monorepo-example/pull/589) [`bbc1a8f`](https://github.com/belgattitude/nextjs-monorepo-example/commit/bbc1a8f07500d13ddf3e86f2cb4111f4f22ddb11) Thanks [@belgattitude](https://github.com/belgattitude)! - Add initial example for prisma cli e2e tests
+
+## 2.6.0
+
+### Minor Changes
+
+- [#553](https://github.com/belgattitude/nextjs-monorepo-example/pull/553) [`77e758b`](https://github.com/belgattitude/nextjs-monorepo-example/commit/77e758bbed1bc4f13b99cdd0ed90fa11fde9518f) Thanks [@belgattitude](https://github.com/belgattitude)! - Add eslint-plugin-regexp
+
 ## 2.5.0
 
 ### Minor Changes